// react-files/template/src/core/models/interface.tsx
export interface ProviderEarningsAdmin {
  [x: string]: string | number;
  '#': string;
  service: string;
  user: string;
  paymentType: string;
  amount: string;
  status: string;
  date: string;
  img: string;
}

export interface ProviderEarningsadmindatas {
  ProviderEarningsAdmin: string;
  '#': string;
  service: string;
  user: string;
  paymentType: string;
  amount: string;
  status: string;
  date: string;
  img: string;
}
export interface AppState {
  mouseOverSidebar: string;
  toggleSidebar2: string;
  toggleSidebar: boolean;
  mobileSidebar: boolean;
}

export interface Pageinfos {
  current_route: string;
  base: string;
  page: string;
  last: string;
}
// ...
export interface CurrentRoute {
  current_route_array: string;
  id: number;
  path: string;
}